+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.0.4
+
+* [Functionality] Adding the possibility to deal with nan values in some channels.
+* [Bug] All events in file sessions are now transformed to the dataframe. See issue #1
+
+## v0.0.3
+
+* Transform data session to dataframe with multiple channels
+
 ## v0.0.2
 
 * Adding documentation